Exit a list in adding/editing tasks cancels the edit mode.
Before TaskTableController changes content, cancel all rows we are
editing. This will in turn re-enable the "Add Task" if it was in a
disabled state (if user where in the middle of adding a task).

Maybe we should ask the user if he/she wants to actually leave if they
are editing tasks?