User are now able to regret a delete action.
Don't think this is 100% flawless, as we can run
in to problem in the timespan when the record
is in a deleted state, ember for some reason reloads
all records from server and tries to load data.

I have not encountered it in the current UI as the
code is now, but I did encounter it when I
transitioned to 'index' instead of 'lists'. Ember
then reloaded all lists, and BOOM we where kinda
fucked.